requirements.txt
bidict==0.23.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
h11==0.14.0
pycryptodome==3.20.0
pymongo==4.6.3
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
wsproto==1.2.0
